backend/requirements.txt
# Core FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10  # PostgreSQL driver

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# Environment Variables
python-dotenv==1.0.1

# Image Processing & AI (Install nanti kalau sudah butuh)
# pillow-simd==10.4.0
anthropic==0.39.0

# HTTP Requests (untuk API external seperti price data)
httpx==0.28.1
requests==2.32.3

# Data Validation - Versi yang kompatibel dengan Python 3.14
pydantic==2.8.2
pydantic-settings==2.3.4

# Date & Time
python-dateutil==2.9.0.post0

# CORS
fastapi-cors==0.0.6

# Testing (Optional tapi recommended)
pytest==8.3.4
pytest-asyncio==0.24.0

# Code Quality (Optional)
black==24.10.0
flake8==7.1.1